+++ src/sys/mem/mod.rs
@@ -16,7 +16,6 @@
 
 mod alloc; // Alloc, ::alloc::alloc::*
 mod borrow; // Mow
-mod cache_align; // CacheAlign
 mod ext; // ExtMem
 mod pin; // Pinned, ::core::pin::*
 mod ptr; // Ptr, ::core::ptr::*
@@ -39,8 +38,15 @@ pub mod cell; // ExtCellOption, ::core::cell::*
 crate::structural_mods! { // _mods, _pub_mods, _hidden
     _mods {
         pub use super::{
-            alloc::_all::*, borrow::_all::*, cache_align::*, ext::*, pin::_all::*,
-            ptr::_all::*, reexports::*, size::_all::*, slice::_all::*, storage::*,
+            alloc::_all::*,
+            borrow::_all::*,
+            ext::*,
+            pin::_all::*,
+            ptr::_all::*,
+            reexports::*,
+            size::_all::*,
+            slice::_all::*,
+            storage::*,
         };
         #[cfg(all(not(feature = "safe_mem"), feature = "unsafe_layout"))]
         pub use super::pod::MemPod;